+++ app/src/Diskora.Core/Models/VirtualDiskFormat.cs
@@ -6,4 +6,5 @@ public enum VirtualDiskFormat
     Vhd,
     Vhdx,
     Iso,
+    Img,
 }